requirements.txt
Adafruit-DHT==1.4.0
attrs==21.4.0
certifi==2020.6.20
chardet==4.0.0
colorzero==1.1
cryptography==3.3.2
dbus-python==1.2.16
distlib==0.3.5
distro==1.5.0
dnspython==2.0.0
filelock==3.7.1
gpg===1.14.0-unknown
gpiozero==1.6.2
grafanalib==0.6.3
idna==2.10
influxdb==5.3.1
influxdb-client==1.30.0
Markdown==3.3.4
mfrc522==0.0.7
msgpack==1.0.4
numpy==1.19.5
olefile==0.46
pexpect==4.8.0
Pillow==8.1.2
pimoroni-sgp30==0.0.2
platformdirs==2.5.2
psutil==5.9.1
Pygments==2.7.1
PyGObject==3.38.0
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.3.1
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==0.9.1
RPi.GPIO==0.7.0
Rx==3.2.0
six==1.16.0
smbus==1.1.post2
smbus2==0.2.1
SPI-Py==1.0
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
urllib3==1.26.5
virtualenv==20.15.1
yarg==0.1.9
